--- a/multibody/tree/body_node.h
+++ b/multibody/tree/body_node.h
@@ -242,6 +242,10 @@ class BodyNode : public MultibodyElement<T> {
       const FrameBodyPoseCache<T>& frame_body_pose_cache, const T* positions,
       PositionKinematicsCache<T>* pc) const = 0;
 
+  virtual void CalcPositionKinematicsCacheInM_BaseToTip(
+      const FrameBodyPoseCache<T>& frame_body_pose_cache, const T* positions,
+      PositionKinematicsCacheInM<T>* pcm) const = 0;
+
   // Calculates the hinge matrix H_PB_W, the `6 x nm` hinge matrix that relates
   // V_PB_W`(body B's spatial velocity in its parent body P, expressed in world
   // W) to this node's nm generalized velocities (or mobilities) v_B as
@@ -291,6 +295,10 @@ class BodyNode : public MultibodyElement<T> {
       const std::vector<Vector6<T>>& H_PB_W_cache, const T* velocities,
       VelocityKinematicsCache<T>* vc) const = 0;
 
+  virtual void CalcVelocityKinematicsCacheInM_BaseToTip(
+      const T* positions, const PositionKinematicsCacheInM<T>& pcm,
+      const T* velocities, VelocityKinematicsCacheInM<T>* vcm) const = 0;
+
   // The CalcMassMatrix() algorithm invokes this on each body k, serving
   // as the composite body R(k) in the outer loop of Jain's algorithm 9.3.
   // This node must fill in its nv x nv diagonal block in M, and then
@@ -301,6 +309,12 @@ class BodyNode : public MultibodyElement<T> {
       const PositionKinematicsCache<T>& pc,
       const std::vector<SpatialInertia<T>>& K_BBo_W_cache,  // composites
       const std::vector<Vector6<T>>& H_PB_W_cache,
+      EigenPtr<MatrixX<T>> M) const = 0;
+
+  // Same, but calculated in body M frames rather than World.
+  virtual void CalcMassMatrixContributionViaM_TipToBase(
+      const T* positions, const PositionKinematicsCacheInM<T>& pcm,
+      const std::vector<SpatialInertia<T>>& I_BMo_M_cache,
       EigenPtr<MatrixX<T>> M) const = 0;
 
   // There are six functions for calculating the off-diagonal blocks, one for
@@ -323,6 +337,22 @@ class BodyNode : public MultibodyElement<T> {
   DECLARE_MASS_MATRIX_OFF_DIAGONAL_BLOCK_VIA_WORLD(6);
 
 #undef DECLARE_MASS_MATRIX_OFF_DIAGONAL_BLOCK_VIA_WORLD
+
+  // Repeat for the M-frame algorithm.
+#define DECLARE_MASS_MATRIX_OFF_DIAGONAL_BLOCK_VIA_M(Bnv)           \
+  virtual void CalcMassMatrixOffDiagonalBlockViaM##Bnv(             \
+      const T* positions, const PositionKinematicsCacheInM<T>& pcm, \
+      int B_start_in_v, const Eigen::Matrix<T, 6, Bnv>& Fm_CMp_Mp,  \
+      EigenPtr<MatrixX<T>> M) const = 0
+
+  DECLARE_MASS_MATRIX_OFF_DIAGONAL_BLOCK_VIA_M(1);
+  DECLARE_MASS_MATRIX_OFF_DIAGONAL_BLOCK_VIA_M(2);
+  DECLARE_MASS_MATRIX_OFF_DIAGONAL_BLOCK_VIA_M(3);
+  DECLARE_MASS_MATRIX_OFF_DIAGONAL_BLOCK_VIA_M(4);
+  DECLARE_MASS_MATRIX_OFF_DIAGONAL_BLOCK_VIA_M(5);
+  DECLARE_MASS_MATRIX_OFF_DIAGONAL_BLOCK_VIA_M(6);
+
+#undef DECLARE_MASS_MATRIX_OFF_DIAGONAL_BLOCK_VIA_M
 
   // This method is used by MultibodyTree within a base-to-tip loop to compute
   // this node's kinematics that depend on the generalized accelerations, i.e.
@@ -374,6 +404,12 @@ class BodyNode : public MultibodyElement<T> {
       const VelocityKinematicsCache<T>* vc, const T* accelerations,
       std::vector<SpatialAcceleration<T>>* A_WB_array) const = 0;
 
+  virtual void CalcSpatialAccelerationInM_BaseToTip(
+      const T* positions, const PositionKinematicsCacheInM<T>& pcm,
+      const T* velocities, const VelocityKinematicsCacheInM<T>& vcm,
+      const T* accelerations,
+      std::vector<SpatialAcceleration<T>>* A_WM_M_array) const = 0;
+
   // Computes the generalized forces `tau` for a single BodyNode.
   // This method is used by MultibodyTree within a tip-to-base loop to compute
   // the vector of generalized forces `tau` that would correspond with a known
@@ -416,13 +452,13 @@ class BodyNode : public MultibodyElement<T> {
   //   result in body B having spatial acceleration `A_WB`. This can be the same
   //   object as tau_applied_array.
   //
-  // @pre CalcInverseDynamics_TipToBase() must have already been
+  // @pre CalcInverseDynamicsViaWorld_TipToBase() must have already been
   // called for all the child nodes of `this` node (and, by recursive
   // precondition, all successor nodes in the tree.)
   // Unit test coverage for this method is provided, among others, in
   // double_pendulum_test.cc, and by any other unit tests making use of
   // MultibodyTree::CalcInverseDynamics().
-  virtual void CalcInverseDynamics_TipToBase(
+  virtual void CalcInverseDynamicsViaWorld_TipToBase(
       const FrameBodyPoseCache<T>& frame_body_pose_cache, const T* positions,
       const PositionKinematicsCache<T>& pc,
       const std::vector<SpatialInertia<T>>& M_B_W_cache,
@@ -431,6 +467,17 @@ class BodyNode : public MultibodyElement<T> {
       const std::vector<SpatialForce<T>>& Fapplied_Bo_W_array,
       const Eigen::Ref<const VectorX<T>>& tau_applied_array,
       std::vector<SpatialForce<T>>* F_BMo_W_array,
+      EigenPtr<VectorX<T>> tau_array) const = 0;
+
+  virtual void CalcInverseDynamicsViaM_TipToBase(
+      const FrameBodyPoseCache<T>& frame_body_pose_cache,  // M_BMo_M, X_BM
+      const T* positions,
+      const PositionKinematicsCacheInM<T>& pc,  // X_MpM, X_WB
+      const VelocityKinematicsCacheInM<T>& vc,  // V_WM_M
+      const std::vector<SpatialAcceleration<T>>& A_WM_M_array,
+      const std::vector<SpatialForce<T>>& Fapplied_Bo_W_array,  // Bo, W !
+      const Eigen::Ref<const VectorX<T>>& tau_applied_array,
+      std::vector<SpatialForce<T>>* F_BMo_M_array,
       EigenPtr<VectorX<T>> tau_array) const = 0;
 
   // This method is used by MultibodyTree within a tip-to-base loop to compute
@@ -618,6 +665,11 @@ class BodyNode : public MultibodyElement<T> {
       const std::vector<SpatialInertia<T>>& M_BBo_W_all,
       std::vector<SpatialInertia<T>>* K_BBo_W_all) const;
 
+  virtual void CalcCompositeBodyInertiaInM_TipToBase(
+      const FrameBodyPoseCache<T>& frame_body_pose_cache,  // for M_BMo_M
+      const PositionKinematicsCacheInM<T>& pcm,            // for X_MpM(q)
+      std::vector<SpatialInertia<T>>* K_BMo_M_all) const;
+
   // Forms LLT factorization of articulated rigid body's hinge inertia matrix.
   // @param[in] D_B Articulated rigid body hinge matrix.
   // @param[out] llt_D_B Stores the LLT factorization of D_B.
@@ -721,6 +773,33 @@ SPECIALIZE_MASS_MATRIX_VIA_WORLD_DISPATCHER(5);
 SPECIALIZE_MASS_MATRIX_VIA_WORLD_DISPATCHER(6);
 
 #undef SPECIALIZE_MASS_MATRIX_VIA_WORLD_DISPATCHER
+
+// Same as above except for CalcMassMatrixViaM().
+template <typename T, int Bnv>
+class CalcMassMatrixOffDiagonalViaMDispatcher;
+
+#define SPECIALIZE_MASS_MATRIX_VIA_M_DISPATCHER(Bnv)                         \
+  template <typename T>                                                      \
+  class CalcMassMatrixOffDiagonalViaMDispatcher<T, Bnv> {                    \
+   public:                                                                   \
+    static void Dispatch(const BodyNode<T>& parent_node, const T* positions, \
+                         const PositionKinematicsCacheInM<T>& pcm,           \
+                         int B_start_in_v,                                   \
+                         const Eigen::Matrix<T, 6, Bnv>& Fm_CMp_Mp,          \
+                         EigenPtr<MatrixX<T>> M) {                           \
+      parent_node.CalcMassMatrixOffDiagonalBlockViaM##Bnv(                   \
+          positions, pcm, B_start_in_v, Fm_CMp_Mp, M);                       \
+    }                                                                        \
+  }
+
+SPECIALIZE_MASS_MATRIX_VIA_M_DISPATCHER(1);
+SPECIALIZE_MASS_MATRIX_VIA_M_DISPATCHER(2);
+SPECIALIZE_MASS_MATRIX_VIA_M_DISPATCHER(3);
+SPECIALIZE_MASS_MATRIX_VIA_M_DISPATCHER(4);
+SPECIALIZE_MASS_MATRIX_VIA_M_DISPATCHER(5);
+SPECIALIZE_MASS_MATRIX_VIA_M_DISPATCHER(6);
+
+#undef SPECIALIZE_MASS_MATRIX_VIA_M_DISPATCHER
 
 }  // namespace internal
 }  // namespace multibody