@sherm1 2024-12-05
Preliminary performance results for M-frame Inverse Dynamics. TL/DR: switching from W to M frame resulted in a 25% speedup for ID. Combined with previous changes ID is 2X faster than when we started. Details:

|                |  This PR |  Master  |  Before speedup work
|-----------|----------|-----------|--------------------------
| time μs   |    11.6    |     15.2    |      21.4
| speedup |    24%   |      26%    |      --
| total       |    46%   |      26%    |      --

Cassie benchmark times on my old Puget Xeon E5-2650v4@2.2GHz

I'm still studying this to see where we can squeeze out more speed.
cc @amcastro-tri