@sherm1 2025-03-28
This PR also contains the WIP on optimal F&M frames for revolute joints. With that mostly in and using the specialized axial rotation functions in #22814, just to see where we are I timed Cassie Position Kinematics (on my old Puget with some other things running). This is the existing F-frame algorithm, not the M-frame kinematics which is also in this WIP PR.
|                   |      start of project    | master    |     this PR      | 
|-------------|-----------------------|-----------|----------------
|   time μs    |            5.33               |    3.43     |        2.65       | 
| vs. master |           ----                 |     ----      |        23%       |
| vs. start     |           ----                 |     36%     |       50%       |

Considering there is not much going on in Position Kinematics, IMO a 2X speedup for the original algorithm is a good start! More to come ...