@sherm1 2025-03-31
For comparison (and apples-to-apples times vs published results), times for this PR on my laptop (Xeon W-11855M CPU @ 3.20GHz):
|                   |     Old Puget (gcc)   |  laptop gcc   |  laptop clang   | 
|-------------|-----------------------|--------------|------------------
|   time μs    |            2.65               |       1.06       |         0.975        | 

Caveat: it is hard to get repeatable times on my Ubuntu VM. The above times are the best of many runs.