@sherm1 2025-06-05
Bonus chart: how much better is clang 14.0.0 than gcc 11.4.0? TL;DR: a lot!
See previous post for legend and CPU info.

|                            |     P    |  P+CBI | P+CBI+MM |  P + V | P + V + ID | P + V + FD
|-------------------|--------|---------|--------------|---------|-------------|------------
|   gcc W-frame    |   2.17  |    6.14  |       11.2       |    4.33   |     12.8       | 40.7
|  clang W-frame  |   2.04  |   3.51   |       6.31       |    3.57   |     7.64       | 36.5
|     speedup W     |   6%   |    43%  |        44%      |   18%   |      40%      |  10%
|                            |           |             |                     |              |                  |
|    gcc M-frame   |  1.97   |    5.56  |        9.59      |    3.85   |     10.2       |  --
|  clang M-frame  |  1.93   |    3.17  |       5.26       |    2.61   |     5.74       |  --
|     speedup M     |   2%   |    43%  |        45%      |   32%    |      44%      |  n/a